net5.0 target; fixes; use ParserConstruction namespace for Superpower types
