+++ src/Serilog.Expressions/Expressions/Parsing/ExpressionTokenizer.cs
@@ -1,7 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using Serilog.Superpower;
-using Serilog.Superpower.Model;
 
 namespace Serilog.Expressions.Parsing
 {
