+++ src/Serilog.Expressions/Expressions/Parsing/ParserExtensions.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using Serilog.Superpower;
-using Serilog.Superpower.Model;
 
 namespace Serilog.Expressions.Parsing
 {
