+++ src/Serilog.Expressions/ParserConstruction/Model/TokenListParserResult`2.cs
@@ -13,10 +13,10 @@
 // limitations under the License.
 
 using System;
-using Serilog.Superpower.Display;
-using Serilog.Superpower.Util;
+using Serilog.ParserConstruction.Display;
+using Serilog.ParserConstruction.Util;
 
-namespace Serilog.Superpower.Model
+namespace Serilog.ParserConstruction.Model
 {
     /// <summary>
     /// The result of parsing from a token list.
